chore: refactored debug camera and cleaned up cameras_v2
